adds step RoundingStrategies
Using this strategies, rounding the nearest value
divisible by step is possible.
rounding towards zero and away from zero is supported
test case included
javadoc included


git-svn-id: svn://141.76.65.164/kp-salespoint@323 eb53f5e8-9ec5-4bc0-8d13-06c033283886

+++ SalespointFramework/src/org/salespointframework/core/quantity/rounding/AbstractRoundingStrategy.java
@@ -9,10 +9,10 @@ public abstract class AbstractRoundingStrategy implements RoundingStrategy,
 
 	protected int numberOfDigits;
 	protected int roundingDigit;
-	protected int roundingStep;
+	protected BigDecimal roundingStep;
 
 	public AbstractRoundingStrategy(int numberOfDigits, int roundingDigit,
-			int roundingStep) {
+			BigDecimal roundingStep) {
 		this.numberOfDigits = numberOfDigits;
 		this.roundingDigit = roundingDigit;
 		this.roundingStep = roundingStep;
